+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 0.9.2 [Work in progress]
 
+- `fix` Fixed troubles with overriding user search model (dmeroff)
 - `fix` Fixed bug with when model was defined as array in modelMap (dmeroff)
 - `fix` Removed ajax-request flood on login form (thiagotalma)
 - `enh` Added `th` translation (kongoon)